test: test float32

+++ tests/high_level/test_high_level_inplace.py
@@ -24,7 +24,7 @@ import torchopt
 
 
 @helpers.parametrize(
-    dtype=[torch.float64, torch.float32],
+    dtype=[torch.float32, torch.float64],
     lr=[1e-3, 1e-4],
     momentum=[0.0, 0.1],
     nesterov=[False, True],
@@ -66,7 +66,7 @@ def test_sgd(dtype: torch.dtype, lr: float, momentum: float, nesterov: bool) -> 
 
 
 @helpers.parametrize(
-    dtype=[torch.float64, torch.float32],
+    dtype=[torch.float32, torch.float64],
     lr=[1e-3, 1e-4],
     betas=[(0.9, 0.999), (0.95, 0.9995)],
     eps=[1e-8],
@@ -98,7 +98,7 @@ def test_adam(dtype: torch.dtype, lr: float, betas: Tuple[float, float], eps: fl
 
 
 @helpers.parametrize(
-    dtype=[torch.float64, torch.float32],
+    dtype=[torch.float32, torch.float64],
     lr=[1e-3, 1e-4],
     betas=[(0.9, 0.999), (0.95, 0.9995)],
     eps=[1e-8],
@@ -141,7 +141,7 @@ def test_accelerated_adam_cpu(
 
 @pytest.mark.skipif(not torch.cuda.is_available(), reason='No CUDA device available.')
 @helpers.parametrize(
-    dtype=[torch.float64, torch.float32],
+    dtype=[torch.float32, torch.float64],
     lr=[1e-3, 1e-4],
     betas=[(0.9, 0.999), (0.95, 0.9995)],
     eps=[1e-8],
@@ -185,7 +185,7 @@ def test_accelerated_adam_cuda(
 
 
 @helpers.parametrize(
-    dtype=[torch.float64, torch.float32],
+    dtype=[torch.float32, torch.float64],
     lr=[1e-3, 1e-4],
     alpha=[0.9, 0.99],
     eps=[1e-8],
